+++ CHANGES.md
@@ -9,6 +9,8 @@ Version 1.10
   containing quantified constraints.
 * Fix a bug in which `expandType` would not expand type synonyms in the kinds
   of type variable binders in `forall`s.
+* Locally reified class methods, data constructors, and record selectors now
+  quantify kind variables properly.
 * Add more functions which compute free variables.
 
 Version 1.9